git-number: Use perl instead of /bin/sh
The shell version of git-number eventually call either git-id or
git-list, both of which are perl scripts.  Might as well go all the way
perl and combine git-id, git-list and git-number into a single perl
script. This is the first step in that direction.

Also rewriting git-number in perl makes it more portable since we're
removing the dependency on shell.

It's also better in the sense that we're not spawning more processes
than needed.