Provide a way to recognize numbers after two dashes
We used to pass arguments following two-dashes (--) intact. For example
invocations like this is impossible:

    $ alias g='git-number'
    $ g log -- 1 2

where 1 and 2 are numbers corresponding to  files that are no longer
tracked in git, but are still in the working directory.

Work around this by accepting triple dash, which work likes two dashes
but we still recognize the numbers, like so:

    $ g log --- 1 2

which becomes synonymous with

    $ g log -- $(git-id 1 2)

Fixes #46